Suggested Changes

+++ nc-news-fe/src/components/CommentCard.jsx
@@ -4,10 +4,18 @@ import { ToastContainer, toast } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
 
 const CommentCard = ({ comment, setComments }) => {
+
+  const username = "cooljmessy"
+
+  const date = [];
+  for (let i = 0; i < comment.created_at.length; i += 10) {
+    date.push(comment.created_at.slice(i, i + 10));
+  }
+
   const notify = (message) => {
     toast(message, {
       position: "top-center",
-      autoClose: 5000,
+      autoClose: 1000,
       hideProgressBar: false,
       closeOnClick: true,
       pauseOnHover: true,
@@ -18,10 +26,16 @@ const CommentCard = ({ comment, setComments }) => {
   };
 
   const handleDelete = (commentId) => {
+    if(comment.author !== username){
+      notify("Not the author")
+      return;
+    }
     deleteCommentById(comment.comment_id).then(() => {
       notify("Comment Deleted");
       setComments((prevComments) => {
-        return prevComments.filter((comment) => comment.comment_id !== commentId);
+        return prevComments.filter(
+          (comment) => comment.comment_id !== commentId
+        );
       });
     });
   };
@@ -29,8 +43,20 @@ const CommentCard = ({ comment, setComments }) => {
   return (
     <section className="border border-black p-2 m-2">
       <h6>{comment.body}</h6>
+      <div className="d-flex justify-content-around">
+        <p className="fw-bold">{comment.author}</p>
+        <p className="fw-bold  align-items-center">
+          {date[0]}
+        </p>
+      </div>
       <CommentVoting currentComment={comment} />
-      <button onClick={()=>{handleDelete(comment.comment_id)}}>Delete</button>
+      <button
+        onClick={() => {
+          handleDelete(comment.comment_id);
+        }}
+      >
+        Delete
+      </button>
       <ToastContainer
         position="top-center"
         autoClose={5000}
